Add dovecot-gpgit image
This commits adds a dovecot-gpgit image. With this image it is possible
to automatically encrypt incoming e-mail messages with a user GPG public
key. Currently this requires you to manually import the gpg keys to the
vmail user in the image. As the vmail directory is already persisted,
the gpg keys at /var/mail/.gnugp will be persisted, too.